third_party/WebKit/LayoutTests/animations/svg-attribute-composition/svg-transform-composition-list-expected.txt

This is a testharness.js-based test.
PASS This test uses interpolation-test.js. 
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from add [translate(100 10) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (-0.4) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(92 -6) scale(-6 8) rotate(2 0 -18)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 92 , - 6 , 3 , 0 , - 6 , 0 , 0 , 8 , 0 , 0 , 4 , 2 , 1 , 0.03 , - 0.03 , 1 , - 0.63 , - 0.01 " but got "2 , 0 , 1 , 0 , 0 , 1 , 92 , - 6 , 3 , 0 , 6 , 0 , 0 , 8 , 0 , 0 , 4 , 2 , 1 , 0.03 , - 0.03 , 1 , - 0.63 , - 0.01 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from add [translate(100 10) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (0) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(100 10) scale(10 20) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from add [translate(100 10) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (0.2) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(104 18) scale(12 26) rotate(14 30 54)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 104 , 18 , 3 , 0 , 12 , 0 , 0 , 26 , 0 , 0 , 4 , 14 , 0.97 , 0.24 , - 0.24 , 0.97 , 13.95 , - 5.65 " but got "2 , 0 , 1 , 0 , 0 , 1 , 104 , 18 , 3 , 0 , 12 , 0 , 0 , 26 , 0 , 0 , 4 , 14 , 0.97 , 0.24 , - 0.24 , 0.97 , 13.95 , - 5.65 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from add [translate(100 10) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (0.6) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(112 34) scale(16 38) rotate(22 50 102)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 112 , 34 , 3 , 0 , 16 , 0 , 0 , 38 , 0 , 0 , 4 , 22 , 0.93 , 0.37 , - 0.37 , 0.93 , 41.85 , - 11.3 " but got "2 , 0 , 1 , 0 , 0 , 1 , 112 , 34 , 3 , 0 , 16 , 0 , 0 , 38 , 0 , 0 , 4 , 22 , 0.93 , 0.37 , - 0.37 , 0.93 , 41.85 , - 11.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from add [translate(100 10) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (1) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(120 50) scale(20 50) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from add [translate(100 10) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (1.4) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(128 66) scale(24 62) rotate(38 90 198)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 128 , 66 , 3 , 0 , 24 , 0 , 0 , 62 , 0 , 0 , 4 , 38 , 0.79 , 0.62 , - 0.62 , 0.79 , 140.98 , - 13.44 " but got "2 , 0 , 1 , 0 , 0 , 1 , 128 , 66 , 3 , 0 , 24 , 0 , 0 , 62 , 0 , 0 , 4 , 38 , 0.79 , 0.62 , - 0.62 , 0.79 , 140.98 , - 13.44 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from replace [translate(110 20) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (-0.4) is [translate(110 20) scale(10 20) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 106 , 8 , 3 , 0 , 6 , 0 , 0 , 8 , 0 , 0 , 4 , 2 , 1 , 0.03 , - 0.03 , 1 , - 0.63 , - 0.01 "
PASS Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from replace [translate(110 20) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (0) is [translate(110 20) scale(10 20) rotate(10 20 30)] 
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from replace [translate(110 20) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (0.2) is [translate(110 20) scale(10 20) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 112 , 26 , 3 , 0 , 12 , 0 , 0 , 26 , 0 , 0 , 4 , 14 , 0.97 , 0.24 , - 0.24 , 0.97 , 13.95 , - 5.65 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from replace [translate(110 20) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (0.6) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(120 50) scale(20 50) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 116 , 38 , 3 , 0 , 16 , 0 , 0 , 38 , 0 , 0 , 4 , 22 , 0.93 , 0.37 , - 0.37 , 0.93 , 41.85 , - 11.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from replace [translate(110 20) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (1) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(120 50) scale(20 50) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) rotate(10 10 10)] from replace [translate(110 20) scale(10 20) rotate(10 20 30)] to add [translate(120 50) scale(20 50) rotate(30 70 150)] at (1.4) is [translate(10 10) scale(10 10) rotate(10 10 10) translate(120 50) scale(20 50) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 124 , 62 , 3 , 0 , 24 , 0 , 0 , 62 , 0 , 0 , 4 , 38 , 0.79 , 0.62 , - 0.62 , 0.79 , 140.98 , - 13.44 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) rotate(10 20 30)] from neutral to replace [translate(130 60) scale(30 60) rotate(40 80 160)] at (-0.4) is [translate(102,4) scale(2 4) rotate(-2 -4 -22)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 102 , 4 , 3 , 0 , 2 , 0 , 0 , 4 , 0 , 0 , 4 , - 2 , 1 , - 0.03 , 0.03 , 1 , 0.77 , - 0.15 " but got "2 , 0 , 1 , 0 , 0 , 1 , - 52 , - 24 , 3 , 0 , - 12 , 0 , 0 , - 24 , 0 , 0 , 4 , - 16 , 0.96 , - 0.28 , 0.28 , 0.96 , 16.4 , - 11.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) rotate(10 20 30)] from neutral to replace [translate(130 60) scale(30 60) rotate(40 80 160)] at (0) is [translate(110 20) scale(10 20) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) rotate(10 20 30)] from neutral to replace [translate(130 60) scale(30 60) rotate(40 80 160)] at (0.2) is [translate(114,28) scale(14 28) rotate(16 32 56)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 114 , 28 , 3 , 0 , 14 , 0 , 0 , 28 , 0 , 0 , 4 , 16 , 0.96 , 0.28 , - 0.28 , 0.96 , 16.68 , - 6.65 " but got "2 , 0 , 1 , 0 , 0 , 1 , 26 , 12 , 3 , 0 , 6 , 0 , 0 , 12 , 0 , 0 , 4 , 8 , 0.99 , 0.14 , - 0.14 , 0.99 , 4.61 , - 1.92 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) rotate(10 20 30)] from neutral to replace [translate(130 60) scale(30 60) rotate(40 80 160)] at (0.6) is [translate(122,44) scale(22 44) rotate(28 56 108)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 122 , 44 , 3 , 0 , 22 , 0 , 0 , 44 , 0 , 0 , 4 , 28 , 0.88 , 0.47 , - 0.47 , 0.88 , 57.26 , - 13.65 " but got "2 , 0 , 1 , 0 , 0 , 1 , 78 , 36 , 3 , 0 , 18 , 0 , 0 , 36 , 0 , 0 , 4 , 24 , 0.91 , 0.41 , - 0.41 , 0.91 , 43.2 , - 11.22 "
PASS Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) rotate(10 20 30)] from neutral to replace [translate(130 60) scale(30 60) rotate(40 80 160)] at (1) is [translate(130 60) scale(30 60) rotate(40 80 160)] 
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) rotate(10 20 30)] from neutral to replace [translate(130 60) scale(30 60) rotate(40 80 160)] at (1.4) is [translate(138,76) scale(38 76) rotate(52 104 212)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 138 , 76 , 3 , 0 , 38 , 0 , 0 , 76 , 0 , 0 , 4 , 52 , 0.62 , 0.79 , - 0.79 , 0.62 , 207.03 , - 0.47 " but got "2 , 0 , 1 , 0 , 0 , 1 , 182 , 84 , 3 , 0 , 42 , 0 , 0 , 84 , 0 , 0 , 4 , 56 , 0.56 , 0.83 , - 0.83 , 0.56 , 235.07 , 5.89 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) skewY(10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (-0.4) is [translate(10 10) scale(10 10) skewX(10) skewY(10) translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) skewY(10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0) is [translate(10 10) scale(10 10) skewX(10) skewY(10) translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) skewY(10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0.2) is [translate(10 10) scale(10 10) skewX(10) skewY(10) translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) skewY(10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0.6) is [translate(10 10) scale(10 10) skewX(10) skewY(10) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) skewY(10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (1) is [translate(10 10) scale(10 10) skewX(10) skewY(10) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) skewY(10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (1.4) is [translate(10 10) scale(10 10) skewX(10) skewY(10) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (-0.4) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0.2) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0.6) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (1) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(100 10) scale(0 10) skewX(0) rotate(0 10 20)] to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (1.4) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] at (-0.4) is [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10) translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] at (0) is [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10) translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] at (0.2) is [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10) translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] at (0.6) is [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10) translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 6 , - 40 , 1 , - 0.84 , 0 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 6 , - 40 , 1 , - 0.84 , 0 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] at (1) is [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10) translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 6 , - 40 , 1 , - 0.84 , 0 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 6 , - 40 , 1 , - 0.84 , 0 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] to add [translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] at (1.4) is [translate(10 10) scale(10 10) skewY(10) rotate(10 10 10) translate(120 50) scale(20 50) skewY(-40) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 6 , - 40 , 1 , - 0.84 , 0 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 6 , - 40 , 1 , - 0.84 , 0 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] from neutral to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (-0.4) is [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] from neutral to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0) is [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] from neutral to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0.2) is [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] from neutral to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (0.6) is [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] from neutral to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (1) is [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30)] from neutral to add [translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] at (1.4) is [translate(110 20) scale(10 20) skewX(10) rotate(10 20 30) translate(130 60) scale(30 60) skewY(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 6 , - 30 , 1 , - 0.58 , 0 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (-0.4) is [translate(10 10) scale(10 10) translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0) is [translate(10 10) scale(10 10) translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0.2) is [translate(10 10) scale(10 10) translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0.6) is [translate(10 10) scale(10 10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (1) is [translate(10 10) scale(10 10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (1.4) is [translate(10 10) scale(10 10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10)] from add [translate(100 10) scale(0 10) skewX(0)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (-0.4) is [translate(10 10) scale(10 10) skewX(10) translate(100 10) scale(0 10) skewX(0)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10)] from add [translate(100 10) scale(0 10) skewX(0)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0) is [translate(10 10) scale(10 10) skewX(10) translate(100 10) scale(0 10) skewX(0)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10)] from add [translate(100 10) scale(0 10) skewX(0)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0.2) is [translate(10 10) scale(10 10) skewX(10) translate(100 10) scale(0 10) skewX(0)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 100 , 10 , 3 , 0 , 0 , 0 , 0 , 10 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10)] from add [translate(100 10) scale(0 10) skewX(0)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0.6) is [translate(10 10) scale(10 10) skewX(10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10)] from add [translate(100 10) scale(0 10) skewX(0)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (1) is [translate(10 10) scale(10 10) skewX(10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10)] from add [translate(100 10) scale(0 10) skewX(0)] to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (1.4) is [translate(10 10) scale(10 10) skewX(10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] at (-0.4) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] at (0) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] at (0.2) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] at (0.6) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 5 , - 40 , 1 , 0 , - 0.84 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 5 , - 40 , 1 , 0 , - 0.84 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] at (1) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 5 , - 40 , 1 , 0 , - 0.84 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 5 , - 40 , 1 , 0 , - 0.84 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10)] from add [translate(110 20) scale(10 20) skewX(10)] to add [translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] at (1.4) is [translate(10 10) scale(10 10) skewX(10) rotate(10 10 10) translate(120 50) scale(20 50) skewX(-40) rotate(30 70 150)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 5 , - 40 , 1 , 0 , - 0.84 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 120 , 50 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 , 5 , - 40 , 1 , 0 , - 0.84 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 84.38 , - 14.9 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10)] from neutral to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (-0.4) is [translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10)] from neutral to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0) is [translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10)] from neutral to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0.2) is [translate(110 20) scale(10 20) skewX(10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10)] from neutral to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (0.6) is [translate(110 20) scale(10 20) skewX(10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10)] from neutral to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (1) is [translate(110 20) scale(10 20) skewX(10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(110 20) scale(10 20) skewX(10)] from neutral to add [translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] at (1.4) is [translate(110 20) scale(10 20) skewX(10) translate(130 60) scale(30 60) skewX(-30) rotate(40 80 160)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 110 , 20 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "2 , 0 , 1 , 0 , 0 , 1 , 130 , 60 , 3 , 0 , 30 , 0 , 0 , 60 , 0 , 0 , 5 , - 30 , 1 , 0 , - 0.58 , 1 , 0 , 0 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10)] from add [rotate(10 20 30)] to replace [translate(20 20) rotate(10 10 10)] at (-0.4) is [translate(6 6) rotate(10 24 38)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 6 , 6 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 6.96 , - 3.59 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10)] from add [rotate(10 20 30)] to replace [translate(20 20) rotate(10 10 10)] at (0) is [translate(10 10) rotate(10 20 30)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10)] from add [rotate(10 20 30)] to replace [translate(20 20) rotate(10 10 10)] at (0.2) is [translate(12 12) rotate(10 18 26)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 12 , 12 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 4.79 , - 2.73 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10)] from add [rotate(10 20 30)] to replace [translate(20 20) rotate(10 10 10)] at (0.6) is [translate(16 16) rotate(10 14 18)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 16 , 16 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 3.34 , - 2.16 " but got "2 , 0 , 1 , 0 , 0 , 1 , 20 , 20 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 "
PASS Web Animations: Interpolate attribute <transform> with underlying [translate(10 10)] from add [rotate(10 20 30)] to replace [translate(20 20) rotate(10 10 10)] at (1) is [translate(20 20) rotate(10 10 10)] 
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10)] from add [rotate(10 20 30)] to replace [translate(20 20) rotate(10 10 10)] at (1.4) is [translate(24 24) rotate(10 6 2)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 24 , 24 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 0.44 , - 1.01 " but got "2 , 0 , 1 , 0 , 0 , 1 , 20 , 20 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 20)] from add [rotate(10 20 30) translate(5 15)] to replace [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] at (-0.4) is [translate(6 6) scale(6 16) rotate(10 24 38) translate(5 19)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 6 , 6 , 3 , 0 , 6 , 0 , 0 , 16 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 6.96 , - 3.59 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 19 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 20)] from add [rotate(10 20 30) translate(5 15)] to replace [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] at (0) is [translate(10 10) scale(10 20) rotate(10 20 30) translate(5 15)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 3 , 0 , 10 , 0 , 0 , 20 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 20)] from add [rotate(10 20 30) translate(5 15)] to replace [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] at (0.2) is [translate(12 12) scale(12 22) rotate(10 18 26) translate(5 13)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 12 , 12 , 3 , 0 , 12 , 0 , 0 , 22 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 4.79 , - 2.73 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 13 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 20)] from add [rotate(10 20 30) translate(5 15)] to replace [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] at (0.6) is [translate(16 16) scale(16 26) rotate(10 14 18) translate(5 9)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 16 , 16 , 3 , 0 , 16 , 0 , 0 , 26 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 3.34 , - 2.16 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 9 " but got "2 , 0 , 1 , 0 , 0 , 1 , 20 , 20 , 3 , 0 , 20 , 0 , 0 , 30 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 "
PASS Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 20)] from add [rotate(10 20 30) translate(5 15)] to replace [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] at (1) is [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] 
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(10 10) scale(10 20)] from add [rotate(10 20 30) translate(5 15)] to replace [translate(20 20) scale(20 30) rotate(10 10 10) translate(5 5)] at (1.4) is [translate(24 24) scale(24 34) rotate(10 6 2) translate(5 1)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 24 , 24 , 3 , 0 , 24 , 0 , 0 , 34 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 0.44 , - 1.01 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 1 " but got "2 , 0 , 1 , 0 , 0 , 1 , 20 , 20 , 3 , 0 , 20 , 0 , 0 , 30 , 0 , 0 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 "
Harness: the test ran to completion.
